Game Engine -- added boundingshape to gameobjects

+++ Shared/Shared/Ball.h
@@ -3,7 +3,7 @@
 
 class Ball : public GameObject {
 public:
-	Ball(vec3 position, vec3 velocity, int id, int radius);
+	Ball(vec3 position, vec3 velocity, int id, float radius);
 	GAMEOBJECT_TYPES getGameObjectType() const;
 	void onCollision(GameObject * gameObject);
 	void updateOnServerTick();
